Audio/Video Sync and SyncTest
From class AudioOutput AudioSyncInfo events are sent to class VideoOutput.
The event contains the current time and the PTS of the audio frame
currently processed by the DAC. VideoOutput uses this information to
calculate when to display the next Video frame and starts the timer.

SyncTest: When compiled with macro SYNCTEST defined a test application
is built. It periodically plays the scale of C major (C-Dur-Tonleiter)
with one second and a unique picture for each tone (from dark to light).

+++ timer.hpp
@@ -158,7 +158,7 @@ private:
 
 inline double getSeconds(const timespec_t& t)
 {
-    double d = t.tv_sec + t.tv_nsec / (1000*1000*1000);
+    double d = t.tv_sec + double(t.tv_nsec) / double(1000*1000*1000);
     return d;
 }
